source/requirements.txt
pymongo == 3.12.1
beautifulsoup4 == 4.10.0
flask== 2.0.2
google-api-python-client == 2.30.0
google-auth-httplib2 == 0.1.0
google-auth-oauthlib == 0.4.6
